authorizatoin

+++ controllers/contactController.js
@@ -32,27 +32,35 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = Contact.create({
         name,
         email,
-        phone
+        phone, 
+        user_id: req.user.id,
     });
     res.status(201).json(contact)
 });
 
 const updateContact = asyncHandler(async (req, res) => {
 
-    const contact  = await Contact.findById(req.params.id);
+    try{
+        const contact  = await Contact.findById(req.params.id);
 
-    if(!contact) {
-        res.status(404);
-        throw new Error('contact not found');
-    }
-
-    const updatedContact = await Contact.findByIdAndUpdate(
-        req.params.id,
-        req.body,
-        {new: true}
-    );
+        if(!contact) {
+            res.status(404);
+            throw new Error('contact not found');
+        }
     
-        res.status(200).json(updatedContact);
+        if(contact.user_id.toString() !== req.params.id){
+
+            res.status(403);
+            throw new Error("User doesn't have permission to update other user's cotact.")
+        }
+    
+        const updatedContact = await Contact.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            {new: true}
+        );
+        
+            res.status(200).json(updatedContact);
 
 });
 
@@ -64,6 +72,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if(!contact) {
         res.status(404);
         throw new Error('contact not found');
+    }
+    
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User doesn't have permission to delete other user's cotact.")
     }
 
     await Contact.deleteOne();
